WIP: 217d08b

+++ platform/ps2/ps2.h
@@ -14,7 +14,7 @@ void ps2_SetAudioFormat(unsigned int rate);
 
 unsigned int ps2_pad_read(int port, int slot);
 unsigned int ps2_pad_read_all(void);
-unsigned int ps2_GetTicksInUsec(void);
+u32 ps2_GetTicksInUsec(void);
 void DelayThread(unsigned short int msec);
 
 void ps2_loadHDDModules(void);
@@ -80,8 +80,8 @@ int ps2_remove(const char *file);
 int ps2_mkdir(const char *path, int mode);
 
 struct timeval {
-    long tv_sec;
-    long tv_usec;
+    u32 tv_sec;
+    u32 tv_usec;
 };
 
 #endif
